requirements.txt
Flask
requests
opencv-python
numpy
Pillow
gunicorn
